% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaze_create_surpfiles.R
\name{amaze_create_surpfiles}
\alias{amaze_create_surpfiles}
\title{Create files to determine surprisal of A-maze alternatives}
\usage{
amaze_create_surpfiles(
  x,
  targets = "targets",
  distractors = "distractors",
  position = 4,
  grouping = "Id",
  file = NULL,
  fileinfo = NULL,
  write_to_file = TRUE,
  print_to_console = FALSE
)
}
\arguments{
\item{x}{the df this function runs over.}

\item{targets}{The column with the target stimuli.}

\item{distractors}{The column with the distractor stimuli.}

\item{position}{Surp files are created from Nth position.}

\item{grouping}{The column with the grouping variable (usually sentence ID)}

\item{file}{File path that the files should be written to (if NULL, files are written to the pwd)}

\item{fileinfo}{Additional info to be written to filename (currently not used).}

\item{write_to_file}{Should output be written to file?}

\item{print_to_console}{Should output be written to the console? (Useful for testing).}
}
\value{
Writes separate files for each distractor (from #2).
}
\description{
Runs over a verticalized df with target and distractor stimuli.
For each alternative, prints a file with a sentence of target stimuli
and the distractor as the last word.
Filename will be sentence id and position of the distractor.
}
\examples{
\dontrun{
#'
stims <- tribble(
  ~Id, ~Target, ~Alternative,
  1, "The", "---",
  1, "comment", "watched",
  1, "about", "china",
  1, "his", "cent",
  1, "brother", "imagine",
  1, "made", "push",
  1, "him", "jack",
  1, "feel", "inch",
  1, "uncomfortable.", "qualification.",
  2, "They", "---",
  2, "forced", "senate",
  2, "him", "belt",
  2, "into", "anti",
  2, "going,", "occur,",
  2, "so", "hour",
  2, "he", "jack",
  2, "went.", "lake."
)
amaze_create_surpfiles(stims, targets = "Target", distractors = "Alternative", print_to_console = TRUE)
}
}
