% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.res.R
\name{chisq.res}
\alias{chisq.res}
\title{Determine pearson residuals}
\usage{
chisq.res(x, decimals = 3)
}
\arguments{
\item{x}{Dataframe of tabulated counts (usual input to chisq.test())}

\item{decimals}{Number of decimals in output}
}
\value{
Table of pearson residuals
}
\description{
Function determines pearson residuals from a chi-square test
}
