% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_batch.R
\name{collex_batch}
\alias{collex_batch}
\title{Determine ams for collex() in batch}
\usage{
collex_batch(
  x,
  corpsize = 100000000L,
  ams = NULL,
  str.dir = TRUE,
  decimals = 5,
  cxn.freq = NULL,
  delta.p = TRUE,
  tidy = TRUE
)
}
\arguments{
\item{x}{An input dataframe like for normal collex().}

\item{corpsize}{Size of the corpus (as per normal collex()).}

\item{ams}{A vector with association measures for the batch. If unspecified, returns logl, odds, mi.}

\item{str.dir}{Should directional am be determined? Defaults to \code{TRUE}}

\item{decimals}{Default decimals for the output (as per collex()).}

\item{cxn.freq}{For truncated data sets, full cxn frequency.}

\item{delta.p}{Should deltaP be determined? Defaults to \code{TRUE}}

\item{tidy}{Should the output have lower-case ('tidy') col names?}
}
\value{
A data frame with the usual collex() output. Sorted alphabetically (?)
}
\description{
Function determines multiple association measures at once.
Experimental for now.
}
\examples{
\dontrun{
collex_batch(x)
}
}
