% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detokenize.R
\name{detokenize}
\alias{detokenize}
\title{Detokenize corpus output}
\usage{
detokenize(x, context = context)
}
\arguments{
\item{x}{A dataframe with two columns: cpos and context with the sentences to be detokenized}

\item{context}{A the name of the column with the sentences to be tokenized.}
}
\value{
A data frame with detokenized text.
}
\description{
Output from corpus files in CQP mode are often already tokenized. This function
untokenizes the most basic tokenization issues (and removes some punctuation)
as a quick fix for now. Experimental.
}
\examples{
\dontrun{
detokenize(x)
}
}
