% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detokenize_vec.R
\name{detokenize_vec}
\alias{detokenize_vec}
\title{Detokenize corpus output}
\usage{
detokenize_vec(x, context = context)
}
\arguments{
\item{x}{A dataframe with one cpos and n context columns}

\item{context}{A vector with the names of the cols of sentences to be detokenized.}
}
\value{
A data frame with detokenized text.
}
\description{
Output from corpus files in CQP mode are often already tokenized. This function
untokenizes the most basic tokenization issues (and removes some punctuation)
as a quick fix for now. Experimental. --> this function is to run over
every column of a dataframe
}
\examples{
\dontrun{
detokenize_vec(x)
}
}
