% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrel.R
\name{hrel}
\alias{hrel}
\title{Determine relative entropy (hrel)}
\usage{
hrel(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of length 1 with hrel of the input vector.
}
\description{
Function determines relative entropy (hrel) over a numeric vector
}
\examples{
\dontrun{
hrel(c(0.1, 0.1, 0.1, 0.3, 0.4, 0.8))
}
}
