% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{normalizer}
\alias{normalizer}
\title{Normalize corpus frequencyies}
\usage{
normalizer(x, corpsize, base = "million")
}
\arguments{
\item{x}{Numeric vector of corpus frequencies}

\item{corpsize}{Size of the corpus where the frequency is from.
Default is number of tokens in COCA excl. punctuation}

\item{base}{Base of per x ('million') frequency}
}
\value{
Returns something (update required)
}
\description{
Function to normalize corpus frequencies
}
\examples{
\dontrun{

}
}
