% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cwb.R
\name{read_cwb}
\alias{read_cwb}
\title{Read raw CWB output}
\usage{
read_cwb(path, skip = 11, preserve_tags = FALSE)
}
\arguments{
\item{path}{A filepath to the raw input}

\item{skip}{Number of lines to skip (11 default for cwb header).}

\item{preserve_tags}{If FALSE (default), function will split input into left, match, right}
}
\value{
A tibble with tab-separated left, match, right.
}
\description{
Reads raw CWB output. For now, it can read cpos, left, match, right.
Other output settings TODO.
}
\examples{
\dontrun{
detokenize(x)
}
}
