% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.pcibex.csv.R
\name{read_pcibex_csv}
\alias{read_pcibex_csv}
\title{Read results file from PCIbex Farm}
\usage{
read_pcibex_csv(
  filepath,
  auto.colnames = TRUE,
  fun.col = function(col, cols) {
     cols[cols == col] <- paste(col, "Ibex", sep =
    ".")
     return(cols)
 }
)
}
\arguments{
\item{x}{An input dataframe like for normal collex().}

\item{ams}{A vector with association measures for the batch. If unspecified, returns logl, odds, mi.}

\item{all_ams}{Boolean. If \code{TRUE}, calculates all available ams. Default \code{FALSE}.}

\item{all_combos}{Boolean. If \code{TRUE}, calculate all possible collex combinations. Default \code{FALSE}.}

\item{str.dir}{Should directional am be determined? Defaults to \code{TRUE}}

\item{delta.p}{Should deltaP be determined? Defaults to \code{TRUE}}

\item{decimals}{Default decimals for the output (as per collex()).}

\item{raw}{Default \code{FALSE}, whether input contains raw observations or aggregated counts.}

\item{tidy}{Should the output have lower-case ('tidy') col names? Requires dplyr.}
}
\value{
A data frame with the usual collex.dist() output. Sorted by first association measure.
}
\description{
Function determines multiple association measures at once.
Experimental for now.
}
\examples{
\dontrun{
collex.dist_batch(x)
}
}
