% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suzR.R
\name{theme_suzR}
\alias{theme_suzR}
\title{My theme}
\usage{
theme_suzR(base_family = "IBM Plex Sans")
}
\arguments{
\item{base_family}{Specify the font family name}
}
\value{
Pwetty plot.
}
\description{
Modfiy the theme_minimal(). Refer to: https://ggplot2.tidyverse.org/reference/theme.html.
}
