% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipf.R
\name{zipf}
\alias{zipf}
\title{Determine Zipf}
\usage{
zipf(x, corpsize = 471255965, rounding = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{corpsize}{Size of the corpus where the frequency is from.
Default is number of tokens in COCA excl. punctuation}

\item{rounding}{Desired decimal rounding}
}
\value{
Numeric vector of length 1 with hrel of the input vector.
}
\description{
Function determines a Zipf converted value of corpus frequency
}
