% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb_context.R
\name{cwb_context}
\alias{cwb_context}
\title{Process cwb context sentences}
\usage{
cwb_context(
  x,
  cols = conc,
  delim = "</s>",
  no_of_contexts = 3,
  filter_incomplete = FALSE,
  debug = FALSE
)
}
\arguments{
\item{x}{A dataframe with two columns: cpos and context with the sentences to be detokenized}

\item{cols}{The name of the column with the concordance (usually "conc")}

\item{delim}{The delimiter at which the sentences shall be split.}

\item{no_of_contexts}{An integer with the number of LEFT context sentences}

\item{col_names}{A vector with the column names (default works for 3 context sentences)}
}
\value{
A data frame with detokenized text.
}
\description{
Function to split cwb context
}
\examples{
\dontrun{
process_cwb_context(x)
}
}
