% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb_senlength.R
\name{cwb_senlength}
\alias{cwb_senlength}
\title{Count sentence length in cwb context sentences}
\usage{
cwb_senlength(x, required_length = 3, filter_short = TRUE, remove_cols = TRUE)
}
\arguments{
\item{x}{A dataframe with a cpos column and context sentences}

\item{required_length}{Minimum length of sentences to keep?}

\item{filter_short}{Should the target concs with short context sentences be removed?}

\item{remove_cols}{Should the ancillary columns be removed?}
}
\description{
Function to split cwb context
}
\examples{
\dontrun{
count_sentence_length(x)
}
}
