% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detokenize.R
\name{detokenize}
\alias{detokenize}
\title{Add together two numbers}
\usage{
detokenize(x, context = context)
}
\arguments{
\item{x}{A dataframe with two columns: cpos and context with the sentences to be detokenized}

\item{context}{A the name of the column with the sentences to be tokenized.}
}
\value{
A data frame with detokenized text.
}
\description{
Add together two numbers
}
\examples{
\dontrun{
detokenize(x)
}
}
